fixed dotfile loop to exclude ., .., .DS_*

+++ setup.sh
@@ -171,15 +171,22 @@ git pull
 
 # for dotfile in $(ls -a $INSTALL_TO/me/dotfiles | grep [^.]); do
 for dotfilesrc in $(ls -a $INSTALL_TO/me/dotfiles); do
-	dotfile=".${dotfilesrc}"
-	echo Addding dotfile: $dotfile
 
-	if [ -L $HOME/$dotfile ]; then ### is symbolic link
-		rm $HOME/$dotfile	
-	elif [ -f $HOME/$dotfile ]; then ### is file
-		mv $HOME/$dotfile $HOME/${dotfile}_$(date +"%F_%H.%M.%S")
-	fi
-	ln -s $INSTALL_TO/me/dotfiles/$dotfilesrc $HOME/$dotfile
+    # if [ ${#dotfile} -ge 3 ]; then ### skip . and ..
+    if [[ $dotfilesrc != .* ]]; then ### skip ., .., .DS_*
+		dotfile=".${dotfilesrc}"
+		echo "Addding dotfile: ${dotfile}"
+
+		if [ -L $HOME/$dotfile ]; then ### is symbolic link
+			rm $HOME/$dotfile	
+		elif [ -f $HOME/$dotfile ]; then ### is file
+			mv $HOME/$dotfile $HOME/${dotfile}_$(date +"%F_%H.%M.%S")
+		fi
+		ln -s $INSTALL_TO/me/dotfiles/$dotfilesrc $HOME/$dotfile
+	else
+		echo "Skipping dotfilesrc: ${dotfilesrc}"
+    fi 
+
 done
